Creates PageCleaner module to DRY pagination
